+++ src/Symfony/Component/Security/CHANGELOG.md
@@ -7,6 +7,9 @@ CHANGELOG
 * Removed `Argon2iPasswordEncoder`, use `SodiumPasswordEncoder` instead
 * Removed `BcryptPasswordEncoder`, use `NativePasswordEncoder` instead
 * Removed the `has_role()` function from security expressions, use `is_granted()` instead.
+* `SimpleAuthenticatorInterface`, `SimpleFormAuthenticatorInterface`, `SimplePreAuthenticatorInterface`,
+  `SimpleAuthenticationProvider`, `SimpleAuthenticationHandler`, `SimpleFormAuthenticationListener` and
+  `SimplePreAuthenticationListener` have been removed. Use Guard instead.
 
 4.3.0
 -----